LayoutTests/platform/ios/imported/w3c/web-platform-tests/editing/other/insertparagraph-in-child-of-html.tentative_designMode=on&white-space=normal-expected.txt



FAIL insertParagraph in empty <div style="white-space:normal"> after <body> should split the <div> assert_equals: The <div> should be split expected "<head><title>iframe</title></head><body><br></body><div style=\"white-space:normal\"><br></div><div style=\"white-space:normal\"><br></div>" but got "<head><title>iframe</title></head><body><br></body><div style=\"white-space:normal\"></div>"
FAIL insertParagraph in <div style="white-space:normal"> (containing only a <br>) after <body> should split the <div> assert_equals: The <div> should be split expected "<head><title>iframe</title></head><body><br></body><div style=\"white-space:normal\"><br></div><div style=\"white-space:normal\"><br></div>" but got "<head><title>iframe</title></head><body><br></body><div style=\"white-space:normal\"><br></div>"
FAIL insertParagraph in <div style="white-space:normal"> (containing text) after <body> should not create another <div> element assert_in_array: The <div> should be split value "<head><title>iframe</title></head><body><br></body><div style=\"white-space:normal\">abcd</div>" not in array ["<head><title>iframe</title></head><body><br></body><div style=\"white-space:normal\">ab</div><div style=\"white-space:normal\">cd</div>", "<head><title>iframe</title></head><body><br></body><div style=\"white-space:normal\">ab<br></div><div style=\"white-space:normal\">cd</div>", "<head><title>iframe</title></head><body><br></body><div style=\"white-space:normal\">ab</div><div style=\"white-space:normal\">cd<br></div>", "<head><title>iframe</title></head><body><br></body><div style=\"white-space:normal\">ab<br></div><div style=\"white-space:normal\">cd<br></div>"]
FAIL insertParagraph in empty <div style="white-space:normal"> before <body> should split the <div> assert_equals: The <div> should be split expected "<head><title>iframe</title></head><div style=\"white-space:normal\"><br></div><div style=\"white-space:normal\"><br></div><body><br></body>" but got "<head><title>iframe</title></head><div style=\"white-space:normal\"></div><body><br></body>"
FAIL insertParagraph in <div style="white-space:normal"> (containing only a <br>) before <body> should split the <div> assert_equals: The <div> should be split expected "<head><title>iframe</title></head><div style=\"white-space:normal\"><br></div><div style=\"white-space:normal\"><br></div><body><br></body>" but got "<head><title>iframe</title></head><div style=\"white-space:normal\"><br></div><body><br></body>"
FAIL insertParagraph in <div style="white-space:normal"> (containing text) before <body> should split the <div> assert_in_array: The <div> should be split value "<head><title>iframe</title></head><div style=\"white-space:normal\">abcd</div><body><br></body>" not in array ["<head><title>iframe</title></head><div style=\"white-space:normal\">ab</div><div style=\"white-space:normal\">cd</div><body><br></body>", "<head><title>iframe</title></head><div style=\"white-space:normal\">ab<br></div><div style=\"white-space:normal\">cd</div><body><br></body>", "<head><title>iframe</title></head><div style=\"white-space:normal\">ab</div><div style=\"white-space:normal\">cd<br></div><body><br></body>", "<head><title>iframe</title></head><div style=\"white-space:normal\">ab<br></div><div style=\"white-space:normal\">cd<br></div><body><br></body>"]
FAIL insertParagraph in empty <div style="white-space:normal"> before <head> should split the <div> assert_equals: The <div> should be split expected "<div style=\"white-space:normal\"><br></div><div style=\"white-space:normal\"><br></div><head><title>iframe</title></head><body><br></body>" but got "<div style=\"white-space:normal\"></div><head><title>iframe</title></head><body><br></body>"
FAIL insertParagraph in <div style="white-space:normal"> (containing only a <br>) before <head> should split the <div> assert_equals: The <div> should be split expected "<div style=\"white-space:normal\"><br></div><div style=\"white-space:normal\"><br></div><head><title>iframe</title></head><body><br></body>" but got "<div style=\"white-space:normal\"><br></div><head><title>iframe</title></head><body><br></body>"
FAIL insertParagraph in <div style="white-space:normal"> (containing text) before <head> should split the <div> assert_in_array: The <div> should be split value "<div style=\"white-space:normal\">abcd</div><head><title>iframe</title></head><body><br></body>" not in array ["<div style=\"white-space:normal\">ab</div><div style=\"white-space:normal\">cd</div><head><title>iframe</title></head><body><br></body>", "<div style=\"white-space:normal\">ab<br></div><div style=\"white-space:normal\">cd</div><head><title>iframe</title></head><body><br></body>", "<div style=\"white-space:normal\">ab</div><div style=\"white-space:normal\">cd<br></div><head><title>iframe</title></head><body><br></body>", "<div style=\"white-space:normal\">ab<br></div><div style=\"white-space:normal\">cd<br></div><head><title>iframe</title></head><body><br></body>"]
